+++ R/empty.R
@@ -2,7 +2,7 @@
 #'
 #' Empty items are NULL, NA, or nothing (length 0).
 #' 
-#' @param obj. The object to test.
+#' @param obj ANY. The object to test.
 #' @examples
 #'   is.empty(NULL)
 #'   is.empty(NA)
@@ -15,6 +15,6 @@
 #' @return a boolean whether or not the object is empty.
 #' @export
 is.empty <- function(obj) {
-  if (obj %is% list) { return(all(vapply(obj, is.empty, logical(1)))) }
+  if (methods::is(obj, "list")) { return(all(vapply(obj, is.empty, logical(1)))) }
   suppressWarnings(is.na(obj) || is.null(obj) || NROW(obj) == 0 || obj == "")
 }