+++ R/quickcheck.R
@@ -1,57 +1,56 @@
 #' Create the necessary testing objects to quickcheck a function.
 #' @param fn function. A function to generate test objects for.
-function_test_objects <- ensure(pre = fn %is% "function", post = result %is% list,
-  function(fn) {
-    if (fn %is% validated_function) {
-      preconditions <- preconditions(fn)
-      if (preconditions[[1]] != substitute(list) && is.call(preconditions)) {
-          preconditions <- list(preconditions)
-      }
-      pre_fn <- get_prevalidated_fn(fn)
-      formals <- names(formals(pre_fn))
-      testing_frame <- lapply(seq_along(formals), function(n) sample(test_objects()))
-      testing_frame <- tryCatch(lapply(seq_along(testing_frame), function(pos) {
-        # First we try calculating each input independently so that we can maximize
-        # the number of test samples.
-        frame <- testing_frame[[pos]]
-        Filter(function(item) {
-          env <- list(item)
-          names(env) <- formals[[pos]]
-          for (precondition in as.list(preconditions)) {
-            if (!grepl(formals[[pos]], deparse(precondition), fixed = TRUE)) { next }
-            if (!isTRUE(eval(precondition, env = env))) { return(FALSE) }
-          }
-          TRUE
-        }, frame)
-      }), error = function(e) {
-        if (grepl("not found", as.character(e), fixed = TRUE)) {
-          # If there was a not found error, we assume it was because of interdependent
-          # preconditions, so we go to the backup of calculating the arguments jointly.
-          lapply(lapply(testing_frame, function(frame) {
-            lapply(seq_along(testing_frame[[1]]), function(pos) {
-              env <- lapply(testing_frame, `[[`, pos)
-              names(env) <- formals
-              for (precondition in as.list(preconditions)) {
-                if (!isTRUE(eval(precondition, env = env))) { return(NULL) }
-              }
-              frame[[pos]]
-            })
-          }), function(frame) { Filter(Negate(is.null), frame) })
-        } else { stop(e) }
-      })
-    } else {
-      formals <- names(formals(fn))
-      testing_frame <- lapply(seq_along(formals), function(n) sample(test_objects()))
+function_test_objects <- function(fn) {
+  if (is(fn, "validated_function")) {
+    preconditions <- preconditions(fn)
+    if (preconditions[[1]] != substitute(list) && is.call(preconditions)) {
+        preconditions <- list(preconditions)
     }
-    names(testing_frame) <- formals
-    testing_frame
-  })
+    pre_fn <- checkr::get_prevalidated_fn(fn)
+    formals <- names(formals(pre_fn))
+    testing_frame <- lapply(seq_along(formals), function(n) sample(checkr:::test_objects()))
+    testing_frame <- tryCatch(lapply(seq_along(testing_frame), function(pos) {
+      # First we try calculating each input independently so that we can maximize
+      # the number of test samples.
+      frame <- testing_frame[[pos]]
+      Filter(function(item) {
+        env <- list(item)
+        names(env) <- formals[[pos]]
+        for (precondition in as.list(preconditions)) {
+          if (!grepl(formals[[pos]], deparse(precondition), fixed = TRUE)) { next }
+          if (!isTRUE(eval(precondition, envir = env))) { return(FALSE) }
+        }
+        TRUE
+      }, frame)
+    }), error = function(e) {
+      if (grepl("not found", as.character(e), fixed = TRUE)) {
+        # If there was a not found error, we assume it was because of interdependent
+        # preconditions, so we go to the backup of calculating the arguments jointly.
+        lapply(lapply(testing_frame, function(frame) {
+          lapply(seq_along(testing_frame[[1]]), function(pos) {
+            env <- lapply(testing_frame, `[[`, pos)
+            names(env) <- formals
+            for (precondition in as.list(preconditions)) {
+              if (!isTRUE(eval(precondition, envir = env))) { return(NULL) }
+            }
+            frame[[pos]]
+          })
+        }), function(frame) { Filter(Negate(is.null), frame) })
+      } else { stop(e) }
+    })
+  } else {
+    formals <- names(formals(fn))
+    testing_frame <- lapply(seq_along(formals), function(n) sample(checkr:::test_objects()))
+  }
+  names(testing_frame) <- formals
+  testing_frame
+}
 
 #' Print function arguments
+#' @param x ANY. The object to print args for.
 #' @examples
 #' l <- list(x = seq(3), y = seq(4))
-#' print_args(l)
-#' [1] "x = 1:3, y = 1:4"
+#' checkr:::print_args(l)
 print_args <- function(x) {
   paste0(paste(names(x),
     unname(sapply(x, function(y) {
@@ -63,7 +62,7 @@ print_args <- function(x) {
 
 #' Get the name from a passed function, which may be a validated function or just a block.
 #'
-#' @param orig_function_name. A substituted call of the function.
+#' @param orig_function_name call. A substituted call of the function.
 function_name <- function(orig_function_name) {
   function_name <- if (identical(deparse(as.list(orig_function_name)[[1]]), "ensure")) {
     as.list(orig_function_name)[length(as.list(orig_function_name))][[1]]
@@ -84,17 +83,14 @@ function_name <- function(orig_function_name) {
 #' test the function will be screened ahead of time to ensure they meet the preconditions.
 #'
 #' @param fn function. A function to randomly check postconditions for.
-#' @param postconditions. Optional postconditions to quickcheck for.
+#' @param postconditions list. Optional postconditions to quickcheck for.
 #' @param verbose logical. Whether or not to announce the success.
 #' @param testthat logical. Whether or not to run testthat.
 #' @return either TRUE if the function passed the quickcheck or FALSE if it didn't.
 #' @export
-quickcheck <- ensure(
-  pre = list(fn %is% "function", verbose %is% logical, testthat %is% logical),
-  post = result%is% logical,
-function(fn, postconditions = NULL, verbose = TRUE, testthat = TRUE) {
+quickcheck <- function(fn, postconditions = NULL, verbose = TRUE, testthat = TRUE) {
   post <- substitute(postconditions)
-  testing_frame <- function_test_objects(fn)
+  testing_frame <- checkr:::function_test_objects(fn)
   if (any(vapply(testing_frame, length, numeric(1)) == 0)) {
     stop("No quickcheck testing frame was generated. Make sure your preconditions aren't",
       " impossible to satisfy!")
@@ -106,7 +102,7 @@ function(fn, postconditions = NULL, verbose = TRUE, testthat = TRUE) {
       args <- lapply(testing_frame, `[[`, pos)
       tryCatch({
         result <- do.call(fn, args)
-        validate_(post, env = list(result = result))
+        checkr:::validate_(post, env = list(result = result))
       }, error = function(e) {
         failed <<- TRUE
       })
@@ -122,9 +118,9 @@ function(fn, postconditions = NULL, verbose = TRUE, testthat = TRUE) {
     error_msg <- paste0("Quickcheck for ", function_name, " failed on item #", pos, ": ",
       print_args(args))
     if (isTRUE(verbose) && !isTRUE(testthat)) { message(error_msg) }
-    if (isTRUE(testthat)) { testthat::expect_true(FALSE, error_msg) } 
+    if (isTRUE(testthat)) { testthat::expect_true(FALSE, error_msg) }
     FALSE
   }
-})
+}
 #TODO: Handle splats
 #TODO, but later: Can mix-in your own custom objects into the test objects